requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
black==23.11.0
click==8.1.7
colorama==0.4.6
dill==0.3.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
isort==5.12.0
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==4.0.0
psycopg2==2.9.9
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pylint==3.0.2
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomlkit==0.12.2
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
